# perf(core): only call mkdir for the cache and mirror once