--- a/packages/yarnpkg-core/sources/Cache.ts
+++ b/packages/yarnpkg-core/sources/Cache.ts
@@ -128,10 +128,7 @@ export class Cache {
       await xfs.changeFilePromise(gitignorePath, `/.gitignore\n*.flock\n`);
     }
 
-    const mirrorCwd = this.mirrorCwd;
-    if (mirrorCwd) {
-      await xfs.mkdirPromise(mirrorCwd, {recursive: true});
-    }
+    await xfs.mkdirPromise(this.mirrorCwd || this.cwd, {recursive: true});
   }
 
   async fetchPackageFromCache(locator: Locator, expectedChecksum: string | null, {onHit, onMiss, loader, skipIntegrityCheck}: {onHit?: () => void, onMiss?: () => void, loader?: () => Promise<ZipFS>, skipIntegrityCheck?: boolean}): Promise<[FakeFS<PortablePath>, () => void, string]> {