**What's the problem this PR addresses?**

When populating the cache and/or mirror we call `mkdir` to create the folder(s) before writing every entry which is unnecessary as the cache folder is created during the setup though the folder for the mirror isn't.

**How did you fix it?**

Create the folder for the mirror in `setup` and remove the `mkdir` call from `writeFileWithLock`.

**Result**

Copying from the mirror to the local cache

```
Before
➤ YN0000: ┌ Fetch step
➤ YN0013: │ No packages were cached - 6157 packages had to be fetched
➤ YN0000: └ Completed in 12s 716ms

After
➤ YN0000: ┌ Fetch step
➤ YN0013: │ No packages were cached - 6157 packages had to be fetched
➤ YN0000: └ Completed in 11s 563ms
```

**Checklist**
- [x] I have read the [Contributing Guide](https://yarnpkg.com/advanced/contributing).
- [x] I have set the packages that need to be released for my changes to be effective.
- [x] I will check that all automated PR checks pass before the PR gets reviewed.